Revert commit 0426104
I think zeroconf.decode() only intends to decode additional, unnecessary
backslashes when dots are escaped, i.e. \\\\. decoded as \\.

On the other hand, DNS-SD mandates that <Instance> portions of Service
Instance Names are a single DNS label (RFC6763 section 4.1.1) and recom-
mends that any dots in the <Instance> portion be escaped if concatenating
the three portions of a Service Instance Name (RFC6763 section 4.3).

I therefore consider it a bug when this does not happen and that is what
has been observed in this case. It is thus better not to change the behavior
of zeroconf.decode() but rather to provide a different workaround for
buggy software, if necessary.